Most device status replies
I found that after masking the above code, the group is switched, and the status of the corresponding device can be received. The same operation GenericOnOffSetControllerState can also obtain the status information of most devices. When the number of devices is large, the status of some devices is not available Reply. 

https://github.com/NordicSemiconductor/IOS-nRF-Mesh-Library/issues/43